tests/wpt/web-platform-tests/clipboard-apis/events/paste-event-manual.html

<!doctype html>
<title>The paste event</title>
<link rel="help" href="https://w3c.github.io/clipboard-apis/#clipboard-event-paste">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<input placeholder="Paste any text here to continue" size="100">
<p>Some pre-selected text to copy for convenience</p>
<script>
setup({explicit_timeout: true});
async_test(t => {
  getSelection().selectAllChildren(document.querySelector('p'));
  document.onpaste = t.step_func_done(event => {
    // Nothing can be asserted about the event target until
    // https://github.com/w3c/clipboard-apis/issues/70 is resolved.
    // assert_equals(event.target, document.body, 'event.target');
    assert_true(event.isTrusted, 'event.isTrusted');
    assert_true(event.composed, 'event.composed');
  });
});
</script>
